requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
selenium==4.16.0
webdriver-manager==4.0.1
smtplib-ssl==1.0.0
python-dotenv==1.0.0
lxml==4.9.4
fake-useragent==1.4.0
